+++ extensions/trello/src/utils/postCard.tsx
@@ -11,10 +11,10 @@ export const postCard = async (values: postValues) => {
       idList: values.idList,
       idMembers: values.idMember,
     });
-    showToast({ title: "Success", message: "Your to do was created" });
+    showToast({ title: "Success", message: "Your card was created" });
     popToRoot();
   } catch (error) {
-    showToast(Toast.Style.Failure, "An error occured", "Could not fetch todos, check your credentials");
+    showToast(Toast.Style.Failure, "An error occured", "Could not create card, check your credentials");
     return Promise.resolve([]);
   }
 };